Move tests to test_groupby_subclass.py
