+++ pandas/tests/groupby/test_groupby.py
@@ -195,42 +195,6 @@ def test_inconsistent_return_type():
     tm.assert_series_equal(result, e)
 
 
-@pytest.mark.parametrize(
-    "obj",
-    [
-        tm.SubclassedDataFrame({"A": np.arange(0, 10)}),
-        tm.SubclassedSeries(np.arange(0, 10), name="A"),
-    ],
-)
-def test_groupby_preserves_subclass(obj, groupby_func):
-    # GH28330 -- preserve subclass through groupby operations
-
-    if isinstance(obj, Series) and groupby_func in {"corrwith"}:
-        pytest.skip("Not applicable")
-
-    grouped = obj.groupby(np.arange(0, 10))
-
-    # Groups should preserve subclass type
-    assert isinstance(grouped.get_group(0), type(obj))
-
-    args = []
-    if groupby_func in {"fillna", "nth"}:
-        args.append(0)
-    elif groupby_func == "corrwith":
-        args.append(obj)
-    elif groupby_func == "tshift":
-        args.extend([0, 0])
-
-    result = getattr(grouped, groupby_func)(*args)
-
-    # Reduction or transformation kernels should preserve type
-    slices = groupby_func in {"cumcount", "ngroup", "size"}
-    if isinstance(obj, DataFrame) and slices:
-        assert isinstance(result, type(obj._constructor_sliced(dtype=object)))
-    else:
-        assert isinstance(result, type(obj))
-
-
 def test_pass_args_kwargs(ts, tsframe):
     def f(x, q=None, axis=0):
         return np.percentile(x, q, axis=axis)
